make seperated FakeQueryBuilder Class
